Fix repository naming
The original command to get the git repository name only returned the
base directory name. When we really want is the name of the repository
on the remote server.

We now default to requesting the repo information for `origin` remote.
The fetch URL is filtered down to the following text:

    Interstellar/some-repo